+++ src/lib.rs
@@ -27,6 +27,7 @@
 
 pub mod combined;
 pub mod config;
+pub mod config_reload;
 pub mod pure_redoer;
 pub mod record;
 pub mod redo;